requirements.txt
langchain==0.3.26
langchain-core==0.3.68
langchain-text-splitters==0.3.8
langchain-openai==0.3.5
langchain-community==0.3.27
faiss-cpu==1.11.0
openai==1.63.0
streamlit==1.46.1
matplotlib==3.10.3
seaborn==0.13.2
pdfkit==1.0.0
jinja2==3.1.6
python-dotenv==1.0.1
aiohttp==3.12.13
pandas==2.3.1
numpy>=2.0.0
plotly==5.18.0
pydantic>=2.0.0,<3.0.0 